Added ability to create new layer by defining attributes etc.
