+++ mapstory/static/mapstory/less/base.less
@@ -465,6 +465,44 @@ small {
   margin: 15px auto;
   padding-top: 15px;
 }
-.oauth-body h2{
-  color:#f47a20;
+.oauth-body h2 {
+  color: #f47a20;
+}
+
+/* Layer Create
+----------------------- */
+
+.attribute-table {
+  max-width: 380px;
+  margin-bottom: 0px;
+}
+
+.max-width-200px {
+  max-width: 200px;
+  width: 200px;
+}
+
+.attribute-remove {
+  color: white;
+  border: 0px;
+  padding: 6px 0px 6px 0;
+  margin-top: 4px;
+}
+
+.attribute-remove-icon {
+  color: #0b0b0b;
+  display: block;
+}
+
+.attribute-remove-disabled {
+  cursor: not-allowed;
+}
+
+.attribute-add {
+  color: #0b0b0b;
+  margin-bottom: 20px;
+}
+
+.attribute-add-icon {
+  margin-right: 5px;
 }
